requirements.txt
tqdm==4.64.1
requests==2.28.1
pymssql==2.2.7
pandas==1.5.2
python-dotenv==0.21.0
xmltodict==0.13.0